{"types":"C","runs":24}

+++ Problem25.java
@@ -2,9 +2,9 @@ import acm.program.ConsoleProgram;
 //25. მომხმარებელს შეჰყავს ორი რიცხვი თქვენმა პროგრამამ უნდა დაბეჭდოს ამ ორი რიცხვის უმცირესი საერთო ჯერადი. 
 public class Problem25 extends ConsoleProgram {
 	public void run(){
-		int n = readInt("Give me the first number: ");
-		int m = readInt("Give me the second number: ");
-		int mult = n*m;
+		double n = readInt("Give me the first number: ");
+		double m = readInt("Give me the second number: ");
+		double mult = n*m;
 		for(int i = 0; i < n; i++){
 			if(mult/i % n ==0){
 				if(mult/i % m ==0){
